+++ src/modules/app_commands/preloadView.js
@@ -1,0 +1,15 @@
+/**
+ * Preloads link (src).
+ *
+ * @param {string} src link to preload
+ *
+ * @return {AppCommand}
+ */
+export function preloadView(src) {
+  return {
+    c: 'preloadView',
+    p: {
+      src,
+    },
+  };
+}