+++ src/modules/registerAppEventSubscriptions.js
@@ -1,5 +1,6 @@
 import checkout from './app_event_subscribers/checkout';
 import pageInsetsChanged from './app_event_subscribers/pageInsetsChanged';
+import autologin from './app_event_subscribers/autologin';
 
 /**
  * Subscriptions for app events.
@@ -7,4 +8,5 @@ import pageInsetsChanged from './app_event_subscribers/pageInsetsChanged';
 export default function () {
   pageInsetsChanged();
   checkout();
+  autologin();
 }