+++ src/modules/shopgateApp.js
@@ -1,5 +1,10 @@
 import { getParameterFromQueryString, getQSFromUrl } from './url';
 import { getCookie, setCookie, invalidateCookie } from './cookies';
+import { sendAppCommands } from './sendAppCommands';
+import { presentNotification } from './app_commands/presentNotification';
+import { closeNotification } from './app_commands/closeNotification';
+
+const STORAGE_KEY_LOADING_SCREEN_ENABLED = 'shopgateAppLoadingScreenEnabled';
 
 /**
  * Returns a path app should be redirected when the process on bigcommerce side is done.
@@ -37,4 +42,34 @@ export function checkWebcheckout() {
  */
 export function invalidateWebCheckout() {
   invalidateCookie('sgcloud_checkout');
+}
+
+/**
+ * Shows the loading screen if not already shown.
+ *
+ * @param {number} [timeout=10] For how long the screen will be shown in seconds.
+ */
+export function showLoadingScreen(timeout = 10) {
+  if (window.localStorage.getItem(STORAGE_KEY_LOADING_SCREEN_ENABLED)) {
+    return;
+  }
+  window.localStorage.setItem(STORAGE_KEY_LOADING_SCREEN_ENABLED, 'true');
+
+  sendAppCommands([
+    presentNotification(timeout),
+  ]);
+}
+
+/**
+ * Closes loading screen if one was previously shown.
+ */
+export function closeLoadingScreen() {
+  if (!window.localStorage.getItem(STORAGE_KEY_LOADING_SCREEN_ENABLED)) {
+    return;
+  }
+  window.localStorage.removeItem(STORAGE_KEY_LOADING_SCREEN_ENABLED);
+
+  sendAppCommands([
+    closeNotification(),
+  ]);
 }