+++ src/themes/cornerstone/RegisterSuccess.js
@@ -3,6 +3,7 @@ import { sendAppCommands } from '../../modules/sendAppCommands';
 import { broadcastEvent } from '../../modules/app_commands/broadcastEvent';
 import { getRedirectPath, isWebcheckout, invalidateWebCheckout } from '../../modules/shopgateApp';
 import { redirectToCheckout, prepareForCheckout } from '../../modules/redirectToCheckout';
+import { performAutologin, isAutologinDone } from '../../modules/app_event_subscribers/autologin';
 
 /**
  * Makes register success page escape proof
@@ -19,6 +20,8 @@ export class RegisterSuccess {
    * Makes register success page escape proof
    */
   execute = () => {
+    performAutologin();
+
     this.hideHeader();
     this.hideFooter();
 
@@ -29,8 +32,13 @@ export class RegisterSuccess {
       callback = redirectToCheckout;
     }
     this.rewriteCloseButtonDestination(callback);
-    invalidateWebCheckout();
   };
+
+  /**
+   * Provides information on when the rendering process can be seen as complete.
+   * @returns {boolean}
+   */
+  isRendered = () => isAutologinDone();
 
   /**
    * Sets 'Continue shopping' button close the browser.