+++ CHANGELOG.md
@@ -6,9 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## Unreleased
 
+[0.2.0] - 2018-05-09
+### Added
+- autologin support for user registration on themes: cornerstone and fortune
+- user registration process support on themes: cornerstone and fortune 
+
 [0.1.0] - 2018-02-16
 ### Added
 - theme files for making the checkout escape proof for themes: cornerstone and fortune
 - files for shopgate analytics
 
+[0.2.0]: https://github.com/shopgate/bigcommerce-js/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/shopgate/bigcommerce-js/tree/v0.1.0